Separate component registration into separate files
