+++ atcg_lib/src/Scene/Components/AccelerationStructureComponent.cpp
@@ -1,0 +1,6 @@
+#include <Scene/Components/AccelerationStructureComponent.h>
+
+namespace atcg
+{
+
+}
