+++ atcg_lib/src/Scene/Components/CameraComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/CameraComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT(CameraComponent);
+}
