+++ atcg_lib/src/Scene/Components/CustomRenderComponent.cpp
@@ -1,0 +1,6 @@
+#include <Scene/Components/CustomRenderComponent.h>
+
+namespace atcg
+{
+
+}
