+++ atcg_lib/src/Scene/Components/EdgeCylinderRenderComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/EdgeCylinderRenderComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT(EdgeCylinderRenderComponent);
+}
