+++ atcg_lib/src/Scene/Components/EdgeRenderComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/EdgeRenderComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT(EdgeRenderComponent);
+}
