+++ atcg_lib/src/Scene/Components/EditorCameraComponent.cpp
@@ -1,0 +1,5 @@
+#include <Scene/Components/EditorCameraComponent.h>
+
+namespace atcg
+{
+}
