+++ atcg_lib/src/Scene/Components/GeometryComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/GeometryComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT(GeometryComponent);
+}
