+++ atcg_lib/src/Scene/Components/HeterogeneousMediumComponent.cpp
@@ -1,0 +1,9 @@
+#include <Scene/Components/HeterogeneousMediumComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT_DRAW(HeterogeneousMediumComponent);
+ATCG_REGISTER_COMPONENT_STORE(HeterogeneousMediumComponent);
+ATCG_REGISTER_COMPONENT_SERIALIZATION(HeterogeneousMediumComponent);
+}    // namespace atcg
