+++ atcg_lib/src/Scene/Components/HomogeneousMediumComponent.cpp
@@ -1,0 +1,9 @@
+#include <Scene/Components/HomogeneousMediumComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT_DRAW(HomogeneousMediumComponent);
+ATCG_REGISTER_COMPONENT_STORE(HomogeneousMediumComponent);
+ATCG_REGISTER_COMPONENT_SERIALIZATION(HomogeneousMediumComponent);
+}    // namespace atcg
