+++ atcg_lib/src/Scene/Components/IDComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/IDComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT_SERIALIZATION(IDComponent);
+}    // namespace atcg
