+++ atcg_lib/src/Scene/Components/InstanceRenderComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/InstanceRenderComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT(InstanceRenderComponent);
+}    // namespace atcg
