+++ atcg_lib/src/Scene/Components/MeshLightComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/MeshLightComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT(MeshLightComponent);
+}    // namespace atcg
