+++ atcg_lib/src/Scene/Components/MeshRenderComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/MeshRenderComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT(MeshRenderComponent);
+}    // namespace atcg
