+++ atcg_lib/src/Scene/Components/NameComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/NameComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT_SERIALIZATION(NameComponent);
+}    // namespace atcg
