+++ atcg_lib/src/Scene/Components/PointLightComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/PointLightComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT(PointLightComponent);
+}    // namespace atcg
