+++ atcg_lib/src/Scene/Components/PointRenderComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/PointRenderComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT(PointRenderComponent);
+}    // namespace atcg
