+++ atcg_lib/src/Scene/Components/PointSphereRenderComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/PointSphereRenderComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT(PointSphereRenderComponent);
+}    // namespace atcg
