+++ atcg_lib/src/Scene/Components/RenderComponent.cpp
@@ -1,0 +1,5 @@
+#include <Scene/Components/RenderComponent.h>
+
+namespace atcg
+{
+}
