+++ atcg_lib/src/Scene/Components/ScriptComponent.cpp
@@ -1,0 +1,7 @@
+#include <Scene/Components/ScriptComponent.h>
+#include <Scene/ComponentRegistry.h>
+
+namespace atcg
+{
+ATCG_REGISTER_COMPONENT(ScriptComponent);
+}    // namespace atcg
